+++ packages/flusmic/lib/flusmic.dart
@@ -1,4 +1,5 @@
 library flusmic;
 
+export './src/flusmic_error.dart';
 export './src/flusmic_repository.dart';
 export './src/models/flusmic_models.dart';