+++ packages/flusmic/lib/src/models/types/richable/richable.dart
@@ -7,6 +7,10 @@ import 'embed_data/embed_data.dart';
 part 'richable.freezed.dart';
 part 'richable.g.dart';
 
+/// Abstract class for all text
+/// Richables
+abstract class RichableText {}
+
 ///Base class for items in RichText
 ///
 ///`RichableParagraph`, `RichableImage` or a `RichableEmbed`
@@ -15,6 +19,7 @@ class Richable with _$Richable {
   ///RichableHeading1 model
   ///
   ///Represents a heading1 inside RichText
+  @With.fromString('RichableText')
   const factory Richable.heading1({
     required List<Span> spans,
     required String text,
@@ -24,6 +29,7 @@ class Richable with _$Richable {
   ///RichableHeading2 model
   ///
   ///Represents a heading2 inside RichText
+  @With.fromString('RichableText')
   const factory Richable.heading2({
     required List<Span> spans,
     required String text,
@@ -33,6 +39,7 @@ class Richable with _$Richable {
   ///RichableHeading3 model
   ///
   ///Represents a heading3 inside RichText
+  @With.fromString('RichableText')
   const factory Richable.heading3({
     required List<Span> spans,
     required String text,
@@ -42,6 +49,7 @@ class Richable with _$Richable {
   ///RichableHeading4 model
   ///
   ///Represents a heading4 inside RichText
+  @With.fromString('RichableText')
   const factory Richable.heading4({
     required List<Span> spans,
     required String text,
@@ -51,6 +59,7 @@ class Richable with _$Richable {
   ///RichableHeading5 model
   ///
   ///Represents a heading5 inside RichText
+  @With.fromString('RichableText')
   const factory Richable.heading5({
     required List<Span> spans,
     required String text,
@@ -60,6 +69,7 @@ class Richable with _$Richable {
   ///RichableHeading6 model
   ///
   ///Represents a heading6 inside RichText
+  @With.fromString('RichableText')
   const factory Richable.heading6({
     required List<Span> spans,
     required String text,
@@ -69,6 +79,7 @@ class Richable with _$Richable {
   ///RichableParagraph model
   ///
   ///Represents a paragraph inside RichText
+  @With.fromString('RichableText')
   const factory Richable.paragraph({
     required List<Span> spans,
     required String text,